Added more test fixes

+++ app/Game/Kingdoms/Jobs/CapitalCityUnitRequestCancellationMovement.php
@@ -71,7 +71,9 @@ class CapitalCityUnitRequestCancellationMovement implements ShouldQueue
 
         $capitalCityUnitManagement->possiblyCreateKingdomLog($queueData);
 
-        CapitalCityUnitCancellation::where('id', $this->capitalCityCancellationQueueId)->delete();
+        if (! collect($responseData)->contains(fn ($response) => $response['status'] === CapitalCityQueueStatus::CANCELLATION_REJECTED)) {
+            CapitalCityUnitCancellation::where('id', $this->capitalCityCancellationQueueId)->delete();
+        }
 
         event(new UpdateCapitalCityUnitQueueTable($queueData->character));
     }
@@ -121,7 +123,7 @@ class CapitalCityUnitRequestCancellationMovement implements ShouldQueue
 
             if (is_null($unitQueue)) {
 
-                CapitalCityUnitQueue::where('id', $this->capitalCityCancellationQueueId)->update(['status' => CapitalCityQueueStatus::CANCELLATION_REJECTED]);
+                CapitalCityUnitCancellation::where('id', $this->capitalCityCancellationQueueId)->update(['status' => CapitalCityQueueStatus::CANCELLATION_REJECTED]);
 
                 $messages[] = 'Failed to cancel unit recruitment. Seems it must already be done for unit: '.$gameUnit->name;
 
@@ -129,7 +131,10 @@ class CapitalCityUnitRequestCancellationMovement implements ShouldQueue
 
                 event(new UpdateCapitalCityUnitQueueTable($queueData->character));
 
-                return [];
+                return [
+                    'unit_id' => $unitId,
+                    'status' => CapitalCityQueueStatus::CANCELLATION_REJECTED,
+                ];
             }
 
             $unitService->cancelRecruit($unitQueue);
@@ -148,7 +153,6 @@ class CapitalCityUnitRequestCancellationMovement implements ShouldQueue
     {
 
         $responseLookup = collect($responseData)
-            ->reject(fn ($response) => $response['status'] === CapitalCityQueueStatus::CANCELLATION_REJECTED)
             ->pluck('status', 'unit_id')
             ->toArray();
 
@@ -156,7 +160,7 @@ class CapitalCityUnitRequestCancellationMovement implements ShouldQueue
             $gameUnit = GameUnit::where('name', $request['name'])->first();
 
             if (isset($responseLookup[$gameUnit->id])) {
-                $request['secondary_status'] = CapitalCityQueueStatus::CANCELLED;
+                $request['secondary_status'] = $responseLookup[$gameUnit->id];
             }
 
             return $request;
